Extract reusable logic `ConnectForm` to `useRequestAccess`
