+++ projects/js-packages/publicize-components/src/components/services/connect-form.tsx
@@ -1,24 +1,22 @@
-import { Button, useGlobalNotices } from '@automattic/jetpack-components';
+import { Button } from '@automattic/jetpack-components';
+import { useDispatch } from '@wordpress/data';
 import { __, _x } from '@wordpress/i18n';
 import classNames from 'classnames';
 import { useCallback } from 'react';
-import { requestExternalAccess } from '../../utils';
+import { store } from '../../social-store';
+import { KeyringResult } from '../../social-store/types';
 import { SupportedService } from '../services/use-supported-services';
 import styles from './style.module.scss';
+import { useRequestAccess } from './use-request-access';
 
 type ConnectFormProps = {
 	service: SupportedService;
 	isSmall?: boolean;
-	onConfirm: ( data: unknown ) => void;
 	onSubmit?: VoidFunction;
 	displayInputs?: boolean;
-	isMastodonAlreadyConnected?: ( username: string ) => boolean;
 	hasConnections?: boolean;
 	buttonLabel?: string;
 };
-
-const isValidMastodonUsername = ( username: string ) =>
-	/^@?\b([A-Z0-9_]+)@([A-Z0-9.-]+\.[A-Z]{2,})$/gi.test( username );
 
 /**
  * Connect form component
@@ -30,14 +28,22 @@ const isValidMastodonUsername = ( username: string ) =>
 export function ConnectForm( {
 	service,
 	isSmall,
-	onConfirm,
 	onSubmit,
 	displayInputs,
-	isMastodonAlreadyConnected,
 	hasConnections,
 	buttonLabel,
 }: ConnectFormProps ) {
-	const { createErrorNotice } = useGlobalNotices();
+	const { setKeyringResult } = useDispatch( store );
+
+	const onConfirm = useCallback(
+		( result: KeyringResult ) => setKeyringResult( result ),
+		[ setKeyringResult ]
+	);
+
+	const requestAccess = useRequestAccess( {
+		service,
+		onConfirm,
+	} );
 
 	const onSubmitForm = useCallback(
 		( event: React.FormEvent ) => {
@@ -48,43 +54,11 @@ export function ConnectForm( {
 			if ( onSubmit ) {
 				return onSubmit();
 			}
+
 			const formData = new FormData( event.target as HTMLFormElement );
-			const url = new URL( service.connect_URL );
-
-			switch ( service.ID ) {
-				case 'mastodon': {
-					const instance = formData.get( 'instance' ).toString().trim();
-
-					if ( ! isValidMastodonUsername( instance ) ) {
-						createErrorNotice( __( 'Invalid Mastodon username', 'jetpack' ) );
-
-						return;
-					}
-
-					if ( isMastodonAlreadyConnected?.( instance ) ) {
-						createErrorNotice( __( 'This Mastodon account is already connected', 'jetpack' ) );
-
-						return;
-					}
-
-					url.searchParams.set( 'instance', formData.get( 'instance' ) as string );
-					break;
-				}
-
-				default:
-					break;
-			}
-
-			requestExternalAccess( url.toString(), onConfirm );
+			requestAccess( formData );
 		},
-		[
-			createErrorNotice,
-			isMastodonAlreadyConnected,
-			onConfirm,
-			onSubmit,
-			service.ID,
-			service.connect_URL,
-		]
+		[ onSubmit, requestAccess ]
 	);
 
 	return (
